Read session action as EDN map
In particular this means that actions need not be present on blob
generation. For a static client this is not really relevant. However
a client which wants to provide extension points for dynamically
added actions has to run make-blob every time an extension is added
or removed. This is non-trivial business. With this change the blob
update is not necessary anymore.

Read: one compile step removed.